Conflicts:
	core-examples/src/main/java/com/metsci/glimpse/examples/axis/WrappedAxisExample.java
	core/src/main/java/com/metsci/glimpse/gl/util/GLUtils.java
	core/src/main/java/com/metsci/glimpse/painter/group/WrappedPainter.java
	core/src/main/java/com/metsci/glimpse/painter/shape/DynamicPointSetPainter.java
	core/src/main/java/com/metsci/glimpse/support/QuickUtils.java
	core/src/main/java/com/metsci/glimpse/support/shader/line/LineProgram.java
	core/src/main/java/com/metsci/glimpse/support/shader/point/PointArrayColorProgram.java
	core/src/main/java/com/metsci/glimpse/support/shader/point/PointArrayColorSizeProgram.java
	core/src/main/java/com/metsci/glimpse/support/shader/point/PointAttributeColorSizeProgram.java
	core/src/main/java/com/metsci/glimpse/support/shader/point/PointFlatColorProgram.java
	extras-charts/src/main/java/com/metsci/glimpse/charts/bathy/ContourPainter.java
	extras-worldwind/src/main/java/com/metsci/glimpse/worldwind/event/MouseWrapperWorldwind.java

+++ core/src/main/java/com/metsci/glimpse/support/atlas/support/ImageData.java
@@ -68,12 +68,12 @@ public class ImageData
         return delegate.getLocationY( );
     }
 
-    public final int getCenterX( )
+    public final double getCenterX( )
     {
         return delegate.getCenterX( );
     }
 
-    public final int getCenterY( )
+    public final double getCenterY( )
     {
         return delegate.getCenterY( );
     }
